requirements.txt
streamlit
pandas
plotly
folium
streamlit-folium
python-dotenv
requests
